-moved bagging indices utility functions to utility.py -updated example to prevent overfitting -removed unnecessary code
